// src/types/config.ts
export interface ConfigData {
  id?: number;
  key: string;
  value: string;
  description?: string;
}

export interface CodeDetailData {
  id?: number;
  codeId: number;
  name: string;
  description: string;
  value: string;
  lastUpdatedAt?: string;
}

export interface ConfigMetaData {
  screenNames: string[];
  [key: string]: any; // Cho phép thêm các trường khác trong tương lai
}

export interface Config {
  id: number;
  code: string;
  name: string;
  description: string;
  codeType: string;
  metaData: ConfigMetaData;
  data: ConfigData[];
  status: number; // 1: active, 0: inactive
  createdAt?: string;
  updatedAt?: string;
  lastUpdatedAt?: string;
}

export interface ConfigFormData {
  code: string;
  name: string;
  description: string;
  codeType: string;
  metaData: ConfigMetaData;
  data: ConfigData[];
}